requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.1.2
Flask-JWT-Extended==4.4.2
Flask-MySQLdb==1.0.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
passlib==1.7.4
PyJWT==2.4.0
python-dotenv==0.20.0
Werkzeug==2.1.2
zipp==3.8.0
